requirements.txt
discord.py
requests
flask
audioop-lts
python-dotenv
markdown
pytz
numpy